requirements.txt
openai == 2.21.0
python-dotenv == 1.2.2